requirements_nemo.txt
# NeMo model export requirements
# Note: These are heavy dependencies and may take significant time to install

# Core NeMo requirements
nemo_toolkit[all]>=1.20.0
torch>=2.0.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0
numpy>=1.24.0

# Optional but recommended
tensorboard>=2.13.0
wandb>=0.15.0